requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
bitsandbytes==0.39.0
boto3==1.28.17
botocore==1.31.17
braceexpand==0.1.7
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
clip==1.0
cloudpickle==2.2.1
colorama==0.4.6
cycler==0.11.0
decorator==4.4.2
exceptiongroup==1.1.2
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.52.0rc1
gym==0.26.2
gym-notices==0.0.8
habitat==0.2.1
habitat-sim==0.2.1
huggingface-hub==0.15.1
idna==3.4
ifcfg==0.23
imageio==2.25.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
iniconfig==2.0.0
jmespath==1.0.1
joblib==1.3.1
kiwisolver==1.4.2
llvmlite==0.39.1
lmdb==1.4.0
magnum==0.0.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.3.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
moviepy==2.0.0.dev2
msgpack==1.0.5rc1
multidict==6.0.4
networkx==2.6.3
numba==0.56.4
numpy==1.21.5
numpy-quaternion==2022.4.2
oauthlib==3.2.2
objectio==0.2.29
olefile==0.46
openai==0.27.8
opencv-python==4.7.0.68
packaging==23.0
pandas==1.3.5
pickle5==0.0.12
Pillow==8.3.2
pip==22.3.1
pluggy==1.2.0
proglog==0.1.10
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0rc2
pyasn1-modules==0.3.0rc1
pyparsing==3.0.9
pytest==7.4.0
python-dateutil==2.8.2
pytorch-transformers==1.2.0
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
sacremoses==0.0.53
safetensors==0.3.1
scikit-image==0.19.3
scipy==1.7.3
sentencepiece==0.1.99
setuptools==59.5.0
simplejson==3.18.1
six==1.16.0
smmap==3.0.5
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2021.11.2
tokenizers==0.13.3
tomli==2.0.1
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.64.1
transformers==4.30.2
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
webdataset==0.1.40
Werkzeug==2.2.2
wheel==0.37.1
yacs==0.1.8
yarl==1.9.2
zipp==3.12.0
